+++ circle.yml
@@ -1,6 +1,6 @@
-machine: 
+machine:
   node:
-    version: "4.2.6"
+    version: 6.2.1
 test:
   post:
     - "npm run coveralls"
