+++ src/Accounts.spec.js
@@ -3,7 +3,7 @@
 import chai from 'chai';
 import sinonChai from 'sinon-chai';
 import chaiAsPromised from 'chai-as-promised';
-import Accounts from './index';
+import Accounts from './Accounts';
 
 chai.use(sinonChai);
 chai.use(chaiAsPromised);
