+++ src/client.js
@@ -1,4 +1,5 @@
 const packages = ['', '@accounts/rest'];
+
 const client = packages.reduce((prev, curr) => {
   let requiredPackage = prev;
   try {
@@ -7,13 +8,13 @@ const client = packages.reduce((prev, curr) => {
         requiredPackage = require(curr).default;
     }
   } catch (exception) { // eslint-disable-line no-empty
-
+    console.log(exception);
   }
   return requiredPackage;
 }, false);
 
 if (!client) {
-  throw new Error('Please install on a client adapter for accounts.');
+  throw new Error('Please install a client adapter for accounts.');
 }
 
 export default client;
