+++ src/createStore.js
@@ -1,0 +1,15 @@
+import { createStore, applyMiddleware, compose, combineReducers } from 'redux';
+
+export default ({
+  reducers = {},
+  middleware = [],
+  state,
+}) => {
+  createStore(
+    combineReducers(reducers),
+    state,
+    compose(
+      applyMiddleware(...middleware),
+    )
+  );
+};
